dev: added option to select albums you want lyrics from, and also (hopefully) fixed issue with occasional incorrect hints
The issue with the occasional incorrect hints is probably because the given line is the first/last line of the song, and so there can't really be a previous/next line. I initially did control for this but made a couple typos so it didn't work. Should be fixed now, but this is hard to test.